hugo: Added an option in installation of PM to auto install Process or Plugins.
In a folder placed out of the PM root (at same level of Compiled) we can have another folder named autoinstall where we can put process files (*.pm) and plugin files (*.tar).

When the installation of PM starts (where the user sets the mysql credentials, database name and soon the workspace name) if there are Process or Plugins it will show those to be installed. Once the process of installation ends the webspace created will have the processes and the plugins installed and runnig.